Address deep link review feedback
