--- packages/apps/browser-app/src/business-logic/navigation/Route.ts
@@ -1,2 +1,0 @@
-export type { Route as default } from "@superego/routing";
-export { CollectionRouteView, RouteName } from "@superego/routing";
