--- packages/apps/browser-app/src/business-logic/navigation/RouteUtils.ts
@@ -1,1 +1,0 @@
-export { fromHref, toHref, tryFromHref } from "@superego/routing";
