+++ packages/apps/browser-app/src/components/design-system/Link/Link.tsx
@@ -1,7 +1,6 @@
+import { toHref, type Route } from "@superego/routing";
 import type { RefAttributes } from "react";
 import { type LinkProps, Link as LinkRAC } from "react-aria-components";
-import type Route from "../../../business-logic/navigation/Route.js";
-import { toHref } from "../../../business-logic/navigation/RouteUtils.js";
 
 type Props =
   | (Omit<LinkProps, "href"> & { to: Route })
