+++ packages/apps/browser-app/src/components/routes/Conversations/Conversations.tsx
@@ -1,8 +1,8 @@
+import { RouteName } from "@superego/routing";
 import { PiPlus } from "react-icons/pi";
 import { useIntl } from "react-intl";
 import DataLoader from "../../../business-logic/backend/DataLoader.js";
 import { listConversationsQuery } from "../../../business-logic/backend/hooks.js";
-import { RouteName } from "../../../business-logic/navigation/Route.js";
 import Shell from "../../design-system/Shell/Shell.js";
 import ConversationsTable from "../../widgets/ConversationsTable/ConversationsTable.js";
