+++ packages/apps/browser-app/src/components/routes/CreateCollectionAssisted/CreateCollectionAssisted.tsx
@@ -3,10 +3,10 @@ import {
   type InferenceOptions,
   type Message,
 } from "@superego/backend";
+import { RouteName } from "@superego/routing";
 import { PiCode } from "react-icons/pi";
 import { useIntl } from "react-intl";
 import { useStartConversation } from "../../../business-logic/backend/hooks.js";
-import { RouteName } from "../../../business-logic/navigation/Route.js";
 import useNavigationState from "../../../business-logic/navigation/useNavigationState.js";
 import ResultErrors from "../../design-system/ResultErrors/ResultErrors.js";
 import Shell from "../../design-system/Shell/Shell.js";
