+++ packages/apps/browser-app/src/components/widgets/SearchModal/SearchResults.tsx
@@ -1,10 +1,9 @@
 import type { Milliseconds } from "@superego/global-types";
+import { RouteName, toHref } from "@superego/routing";
 import { useEffect, useState } from "react";
 import { ListBox } from "react-aria-components";
 import { FormattedMessage, useIntl } from "react-intl";
 import { useGlobalData } from "../../../business-logic/backend/GlobalData.js";
-import { RouteName } from "../../../business-logic/navigation/Route.js";
-import { toHref } from "../../../business-logic/navigation/RouteUtils.js";
 import type SearchState from "../../../business-logic/search/SearchState.js";
 import SearchType from "../../../business-logic/search/SearchType.js";
 import CollectionUtils from "../../../utils/CollectionUtils.js";
