+++ packages/apps/browser-app/src/components/widgets/UserMessageContentInput/SendRecordButtons.tsx
@@ -1,3 +1,4 @@
+import { RouteName } from "@superego/routing";
 import { Toolbar } from "react-aria-components";
 import {
   PiGear,
@@ -7,7 +8,6 @@ import {
   PiX,
 } from "react-icons/pi";
 import { useIntl } from "react-intl";
-import { RouteName } from "../../../business-logic/navigation/Route.js";
 import IconButton from "../../design-system/IconButton/IconButton.js";
 import IconLink from "../../design-system/IconLink/IconLink.js";
 import * as cs from "./UserMessageContentInput.css.js";
