+++ packages/ui/routing/src/DeepLinkUtils.test.ts
@@ -8,14 +8,12 @@ import {
 import { CollectionRouteView, RouteName } from "./Route.js";
 
 it("creates and parses a document deep link", () => {
-  // Setup
+  // Exercise
   const route = {
     name: RouteName.Document,
     collectionId: Id.generate.collection(),
     documentId: Id.generate.document(),
   } as const;
-
-  // Exercise
   const deepLink = toDeepLink(route);
   const parsedRoute = fromDeepLink(deepLink);
 
@@ -27,15 +25,13 @@ it("creates and parses a document deep link", () => {
 });
 
 it("creates and parses an app view deep link", () => {
-  // Setup
+  // Exercise
   const route = {
     name: RouteName.Collection,
     collectionId: Id.generate.collection(),
     view: CollectionRouteView.App,
     appId: Id.generate.app(),
   } as const;
-
-  // Exercise
   const deepLink = toDeepLink(route);
   const parsedRoute = fromDeepLink(deepLink);
 
@@ -47,10 +43,8 @@ it("creates and parses an app view deep link", () => {
 });
 
 it("parses a manual deep link without an empty authority", () => {
-  // Setup
-  const deepLink = "superego://collections/Collection_abc?view=Table";
-
   // Exercise
+  const deepLink = "superego://collections/Collection_abc?view=Table";
   const href = toHrefFromDeepLink(deepLink);
 
   // Verify
@@ -58,10 +52,8 @@ it("parses a manual deep link without an empty authority", () => {
 });
 
 it("rejects other protocols", () => {
-  // Setup
-  const deepLink = "https://example.com/collections/Collection_abc";
-
   // Exercise
+  const deepLink = "https://example.com/collections/Collection_abc";
   const route = fromDeepLink(deepLink);
 
   // Verify
@@ -69,10 +61,8 @@ it("rejects other protocols", () => {
 });
 
 it("rejects malformed links", () => {
-  // Setup
-  const deepLink = "not a url";
-
   // Exercise
+  const deepLink = "not a url";
   const route = fromDeepLink(deepLink);
 
   // Verify
