+++ packages/ui/routing/src/index.ts
@@ -5,4 +5,4 @@ export {
 } from "./DeepLinkUtils.js";
 export type { default as Route } from "./Route.js";
 export { CollectionRouteView, RouteName } from "./Route.js";
-export { fromHref, toHref, tryFromHref } from "./RouteUtils.js";
+export { fromHref, toHref } from "./RouteUtils.js";
